feat: 열거형(Menu) 파일 분리 및 프로토콜(BindingOptional) 선언
- JuiceMaker와 FruitStore 두 곳에서 옵셔널 바인딩을 해주기 위해 프로토콜(BindingOptional) 채택 후 구현
