+++ JuiceMaker/JuiceMaker/Model/BindingOptional.swift
@@ -1,0 +1,12 @@
+//
+//  BindingOptional.swift
+//  JuiceMaker
+//
+//  Created by 이예은 on 2022/04/27.
+//
+
+import Foundation
+
+protocol BindingOptional {
+    func unwrapOptional(type: Int?) -> Int
+}
